requirements.txt
Pillow==10.4.0
torch==2.3.1+cu121
torchvision==0.19.0
torchvision==0.18.1+cu121
tqdm==4.65.0
